Renamed Numbers module test packages
