+++ Numbers/src/test/kotlin/mathtools/numbers/IntegerMgrTest.kt
@@ -1,4 +1,4 @@
-package data
+package mathtools.numbers
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
