+++ Numbers/src/test/kotlin/mathtools/numbers/primes/BytePrimeCacheTest.kt
@@ -1,7 +1,6 @@
-package data
+package mathtools.numbers.primes
 
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.RepeatedTest
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
